improve native search BROKEN
